Refactoring: account settings page now includes common dashboard header
