+++ src/apps/dashboard-notifications/index.ts
@@ -1,1 +1,1 @@
-export * from './NotificationTab';
+export * from './Notifications';
