+++ src/i18n/de/dashboard-sidebar.json
@@ -6,5 +6,7 @@
   "Documentation": "Dokumentation",
   "Help": "Hilfe",
   "Sign out": "Logout",
-  "No unread notifications": "Keine ungelesenen Nachrichten"
+  "No unread notifications": "Keine ungelesenen Nachrichten",
+  "View": "Anzeigen",
+  "View all": "Alle anzeigen"
 }
