+++ src/i18n/en/dashboard-sidebar.json
@@ -12,5 +12,7 @@
   "Tutorial": "Tutorial",
   "FAQ": "FAQ",
   "Logout": "Logout",
-  "No unread notifications": "No unread notifications"
+  "No unread notifications": "No unread notifications",
+  "View": "View",
+  "View all": "View all"
 }
