+++ src/pages/[lang]/notifications/index.astro
@@ -1,24 +1,16 @@
 ---
-import DashboardLayout from '@layouts/dashboard/DashboardLayout.astro';
+import BaseLayout from '@layouts/BaseLayout.astro';
 import { createSupabaseServerClient } from '@backend/supabaseServerClient';
 import { getTranslations } from '@i18n';
-import NotificationTab from '@apps/dashboard-notifications/NotificationTab';
-import { listMyInvites } from '@backend/crud/notifications';
+import { Notifications } from '@apps/dashboard-notifications';
 
 const i18n = getTranslations(Astro.request, 'notifications');
 
 const supabase = await createSupabaseServerClient(Astro.request, Astro.cookies);
 if (!supabase)
   return Astro.redirect(`/${i18n.lang}/sign-in`);
-
-const { error, data } = await listMyInvites(supabase);
-
-if (error) { 
-  const error = await fetch(`${Astro.url}/500`);
-  return new Response(error.body, { headers: error.headers, status: 500 }); 
-}
 ---
-<DashboardLayout>
+<BaseLayout title="Notifications">
   <header>
     <nav class="breadcrumbs">
       <ol>
@@ -34,10 +26,9 @@ if (error) {
   </header>
 
   <main>
-    <NotificationTab
+    <Notifications
       client:load
-      i18n={i18n}
-      invitationList={data} />
+      i18n={i18n} />
   </main>
 </DashboardLayout>
