+++ src/themes/default/mixins/index.css
@@ -9,3 +9,7 @@
 .w-full {
   width: 100%;
 }
+
+.center {
+  text-align: center;
+}
